requirements.txt
Django==3.2.10
djangorestframework==3.12.4
psycopg2-binary==2.9.3
django-extensions==3.2.1
drf-yasg==1.21.5
black==23.3.0
isort==5.12.0
mypy==1.2.0
coverage==7.2.3
redis==4.5.5
celery==5.3.1
requests~=2.31.0
transitions==0.9.0
